# RemoveParentModulesPlugin performance