--- a/lib/Chunk.js
+++ b/lib/Chunk.js
@@ -257,7 +257,6 @@ class Chunk {
 		module.removeChunk(this);
 		module.addChunk(otherChunk);
 		otherChunk.addModule(module);
-		module.rewriteChunkInReasons(this, [otherChunk]);
 	}
 
 	replaceChunk(oldChunk, newChunk) {