--- a/lib/Module.js
+++ b/lib/Module.js
@@ -37,7 +37,6 @@ class Module extends DependenciesBlock {
 		this.usedExports = null;
 		this.providedExports = null;
 		this._chunks = new SortableSet(undefined, sortById);
-		this._chunksDebugIdent = undefined;
 		this.warnings = [];
 		this.dependenciesWarnings = [];
 		this.errors = [];
@@ -57,7 +56,6 @@ class Module extends DependenciesBlock {
 		this.usedExports = null;
 		this.providedExports = null;
 		this._chunks.clear();
-		this._chunksDebugIdent = undefined;
 		super.disconnect();
 	}
 
@@ -67,7 +65,6 @@ class Module extends DependenciesBlock {
 		this.index2 = null;
 		this.depth = null;
 		this._chunks.clear();
-		this._chunksDebugIdent = undefined;
 		super.unseal();
 	}
 
@@ -77,12 +74,10 @@ class Module extends DependenciesBlock {
 
 	addChunk(chunk) {
 		this._chunks.add(chunk);
-		this._chunksDebugIdent = undefined;
 	}
 
 	removeChunk(chunk) {
 		if(this._chunks.delete(chunk)) {
-			this._chunksDebugIdent = undefined;
 			chunk.removeModule(this);
 			return true;
 		}
@@ -91,24 +86,6 @@ class Module extends DependenciesBlock {
 
 	isInChunk(chunk) {
 		return this._chunks.has(chunk);
-	}
-
-	getChunkIdsIdent() {
-		if(this._chunksDebugIdent !== undefined) return this._chunksDebugIdent;
-		this._chunks.sortWith(sortByDebugId);
-		const chunks = this._chunks;
-		const list = [];
-		for(const chunk of chunks) {
-			const debugId = chunk.debugId;
-
-			if(typeof debugId !== "number") {
-				return this._chunksDebugIdent = null;
-			}
-
-			list.push(debugId);
-		}
-
-		return this._chunksDebugIdent = list.join(",");
 	}
 
 	forEachChunk(fn) {