--- a/lib/Module.js
+++ b/lib/Module.js
@@ -141,12 +141,6 @@ class Module extends DependenciesBlock {
 		return false;
 	}
 
-	rewriteChunkInReasons(oldChunk, newChunks) {
-		for(let i = 0; i < this.reasons.length; i++) {
-			this.reasons[i].rewriteChunks(oldChunk, newChunks);
-		}
-	}
-
 	isUsed(exportName) {
 		if(this.used === null) return exportName;
 		if(!exportName) return !!this.used;