--- a/lib/optimize/EnsureChunkConditionsPlugin.js
+++ b/lib/optimize/EnsureChunkConditionsPlugin.js
@@ -25,7 +25,6 @@ class EnsureChunkConditionsPlugin {
 									newChunks.push(parent);
 								}
 							});
-							module.rewriteChunkInReasons(chunk, newChunks);
 							chunk.removeModule(module);
 							changed = true;
 						}