--- a/lib/optimize/RemoveParentModulesPlugin.js
+++ b/lib/optimize/RemoveParentModulesPlugin.js
@@ -37,21 +37,11 @@ class RemoveParentModulesPlugin {
 					var chunk = chunks[index];
 					if(chunk.parents.length === 0) continue;
 
-					// TODO consider Map when performance has improved https://gist.github.com/sokra/b36098368da7b8f6792fd7c85fca6311
-					var cache = Object.create(null);
 					var modules = chunk.getModules();
 					for(var i = 0; i < modules.length; i++) {
 						var module = modules[i];
 
-						var dId = module.getChunkIdsIdent();
-						var parentChunksWithModule;
-						if(dId === null) {
-							parentChunksWithModule = allHaveModule(chunk.parents, module);
-						} else if(dId in cache) {
-							parentChunksWithModule = cache[dId];
-						} else {
-							parentChunksWithModule = cache[dId] = allHaveModule(chunk.parents, module);
-						}
+						var parentChunksWithModule = allHaveModule(chunk.parents, module);
 						if(parentChunksWithModule) {
 							module.rewriteChunkInReasons(chunk, Array.from(parentChunksWithModule));
 							chunk.removeModule(module);