--- a/lib/optimize/RemoveParentModulesPlugin.js
+++ b/lib/optimize/RemoveParentModulesPlugin.js
@@ -43,7 +43,6 @@ class RemoveParentModulesPlugin {
 
 						var parentChunksWithModule = allHaveModule(chunk.parents, module);
 						if(parentChunksWithModule) {
-							module.rewriteChunkInReasons(chunk, Array.from(parentChunksWithModule));
 							chunk.removeModule(module);
 						}
 					}