--- a/test/ModuleReason.test.js
+++ b/test/ModuleReason.test.js
@@ -30,48 +30,4 @@ describe("ModuleReason", () => {
 			should(myModuleReason.hasChunk(myChunk)).be.true();
 		});
 	});
-
-	describe("rewriteChunks", () => {
-		it("if old chunk is present, it is replaced with new chunks", () => {
-			myModuleReason.module.addChunk(myChunk);
-			myModuleReason.rewriteChunks(myChunk, [myChunk2]);
-
-			should(myModuleReason.hasChunk(myChunk)).be.false();
-			should(myModuleReason.hasChunk(myChunk2)).be.true();
-		});
-
-		it("if old chunk is not present, new chunks are not added", () => {
-			myModuleReason.rewriteChunks(myChunk, [myChunk2]);
-
-			should(myModuleReason.hasChunk(myChunk)).be.false();
-			should(myModuleReason.hasChunk(myChunk2)).be.false();
-		});
-
-		it("if already rewritten chunk is present, it is replaced with new chunks", () => {
-			myModuleReason.module.addChunk(myChunk);
-			myModuleReason.rewriteChunks(myChunk, [myChunk2]);
-			myModuleReason.rewriteChunks(myChunk2, [myChunk]);
-
-			should(myModuleReason.hasChunk(myChunk)).be.true();
-			should(myModuleReason.hasChunk(myChunk2)).be.false();
-		});
-	});
-
-	describe(".chunks", () => {
-		it("is null if no rewrites happen first", () => {
-			should(myModuleReason.chunks).be.Null();
-		});
-
-		it("is null if only invalid rewrites happen first", () => {
-			myModuleReason.rewriteChunks(myChunk, [myChunk2]);
-			should(myModuleReason.chunks).be.Null();
-		});
-
-		it("is an array of chunks if a valid rewrite happens", () => {
-			myModuleReason.module.addChunk(myChunk);
-			myModuleReason.rewriteChunks(myChunk, [myChunk2]);
-
-			should(myModuleReason.chunks).be.eql([myChunk2]);
-		});
-	});
 });