<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**What kind of change does this PR introduce?**
Performance enhancing.

**Did you add tests for your changes?**
No, but I did update some tests related to my changes.

**If relevant, link to documentation update:**
N/A

**Summary**

There were two commits here, both related to performance. This change reduces our build time by roughly ~10%.

**[Module.getChunkIdsIdent] Remove getChunkIdsIdent from Module.**
This function appears to only have been used to generate a cache key inside
RemoveParentModulesPlugin. With the recent change to using sets, it is actually
faster to find parent chunks containing a module than it is to generate this
cache key.

To profile this, I added a wrapper function temporarily in my branch and timed
the duration of the RemoveParetModulesPlugin.

Before this change it took ~6113 ms when building our config.
After this change it takes ~3160 ms against the same config.

**[ModuleReasons] Remove ModuleReason chunk list.**
I didn't see a need for ModuleReason's to track their chunks separate from the
module's list of chunks. Having a single source of truth seems beneficial. This
has a double benefit of speeding up the RemoveParentModulesPlugin.

Before making this change, but including the previous commit: 3160ms
After making this change: 340ms

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
<!-- Try to link to an open issue for more information. -->

**Does this PR introduce a breaking change?**
Yes, but for undocumented features.  `Module.rewriteChunkInReasons`, `Module.getChunkIdsIdent` and `ModuleReasons.rewriteChunks` have all been removed. I couldn't find any documentation on any of these methods, and I believe that they're only used internally by webpack.

<!-- If this PR introduces a breaking change, please describe the impact and a migration path for existing applications. -->

**Other information**