@gdborton 2017-07-25
@sokra @TheLarkInn I can't seem to get this failing test passing even in master... any idea what's going on?

This is the command that I attempted to run after failing with just npm install. I basically tried to mirror the ci steps.
`rm -rf node_modules; npm install -g yarn; yarn install; yarn link webpack; yarn run circleci:test`

The test fails on both my branch as well as master.