@gdborton 2017-07-27
Closing this out, I can't figure out why ModuleReason.chunks should be different than `ModuleReason.module._chunks` or `ModuleReason.dependency.module._chunks`. In my mind only one of these should be the source of truth, but using either breaks tests and I've got no clue as to why. Opening a different PR with just the first commit.