@TheLarkInn 2017-07-28
Gary let's try and get this tackled with some help. Do you have a good email I can have you shoot me?