@mikesherov 2017-07-28
Don't feel bad. When I first converted moduleReason to have a map impl, I tried this very same thing and had failing tests. The reason is due to the lazy caching of modules and the *conditional* removal from of Chunks. The slight difference adds up to real differences at other stages of compilation. 

You can't assume that a reason is the same as the module itself because the reason only counts